2025-08-15_v1.723: não foi testado. serv_menu.c: Menu "Config->Monit" estava sem título (mostrava lixo). Modificações nos nomes das variáveis e na ordem das declarações.
util_menu.h:
Modificações em nomes.
Removidos os tipos TMenuSystemStruct, TMenuStruct e TMenuItemStruct.
Redeclarados os tipos TMenuSystem, TMenu e TMenuItem utilizando declaração antecipada.

ct_estados.c: ajustes de nomes de variáveis para refletir a mudança em em util_menu.h.
base_eeprom.c: type cast.

+++ src/base_eeprom.c
@@ -46,7 +46,8 @@ uint16_t eeprom_le_word(uint8_t end) {
     //Le o byte menos significativo (primeiro byte).
     temp_int = eeprom_read(end);
     //Le o byte mais significativo (segundo byte))
-    temp_int = temp_int + (eeprom_read(end+1)<<8);
+    //temp_int = temp_int + (eeprom_read(end+1)<<8);
+    temp_int = temp_int + (uint16_t)(eeprom_read(end+1)<<8);
     return temp_int;
 }//eeprom_le_word()
